requirements.txt
fastapi>=0.115,<1.0
httpx>=0.27,<1.0
paramiko>=3.4,<4.0
PyYAML>=6.0,<7.0
uvicorn>=0.30,<1.0
websockets>=12,<16
